feat: Placement Center completes the Shard/Segment logic (#566)
* Refine the Create Shard and Segment logic

+++ src/common/metadata-struct/src/journal/segment.rs
@@ -24,7 +24,7 @@ pub struct JournalSegment {
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct JournalSegmentNode {
-    pub node_id: u32,
+    pub node_id: u64,
     pub data_fold: String,
 }
